Data analysis with Python
The objective is to analyze a Netflix dataset by performing data cleaning, exploration, and answering specific questions:

Check for and remove any duplicate records. Identify any null values and visualize them using a heatmap. Find the Show_Id and Director of "House of Cards". Determine and visualize the year with the highest number of releases using a bar graph. Count and visualize the number of Movies and TV Shows using a bar graph. List all movies released in the year 2000. Identify the top 10 directors by the number of contributions to Netflix. Count the number of movies/shows featuring Tom Cruise. List Netflix's different ratings and count how many Movies got 'TV-14' rating in Canada. Identify the country with the highest number of TV Shows. Sort the dataset by the year of release.